Revised File I/O
The initialize_from_file method reads a file line-by-line and loads the
corresponding values for category names, item names, item prices, and
item quantities into the Inventory object that we will use to make
sales while the program is running.  The save_to_file method does the
opposite, writing to the file with the same format.  We can call
save_to_file after each transaction if we wish.